independent namelistfile.py module (#150)
* ref: move namelist to its own module

* fix: forgot to commit module ...

* fix: missing another change in the original commit
