+++ src/sirocco/parsing/yaml_data_models.py
@@ -373,7 +373,7 @@ class ConfigShellTask(ConfigBaseTask, ConfigShellTaskSpecs):
 
 
 @dataclass(kw_only=True)
-class ConfigIconNamelistSpec:
+class ConfigNamelistFileSpec:
     """Class for namelist specifications
 
     - path is the path to the namelist file considered as template
@@ -383,24 +383,24 @@ class ConfigIconNamelistSpec:
     Example:
 
         >>> path = "/some/path/to/icon.nml"
-        >>> nml_info = ConfigIconNamelistSpec(path=Path(path))
+        >>> nml_info = ConfigNamelistFileSpec(path=Path(path))
     """
 
     path: Path = field(repr=False)
 
 
-class ConfigIconNamelist(BaseModel, ConfigIconNamelistSpec):
+class ConfigNamelistFile(BaseModel, ConfigNamelistFileSpec):
     """
     Validated namelist specifications.
 
     Example:
 
         >>> import textwrap
-        >>> from_init = ConfigIconNamelist(
+        >>> from_init = ConfigNamelistFile(
         ...     path="/path/to/some.nml", specs={"block": {"key": "value"}}
         ... )
         >>> from_yml = validate_yaml_content(
-        ...     ConfigIconNamelist,
+        ...     ConfigNamelistFile,
         ...     textwrap.dedent(
         ...         '''
         ...         /path/to/some.nml:
@@ -411,8 +411,8 @@ class ConfigIconNamelist(BaseModel, ConfigIconNamelistSpec):
         ... )
         >>> from_init == from_yml
         True
-        >>> no_spec = ConfigIconNamelist(path="/path/to/some.nml")
-        >>> no_spec_yml = validate_yaml_content(ConfigIconNamelist, "/path/to/some.nml")
+        >>> no_spec = ConfigNamelistFile(path="/path/to/some.nml")
+        >>> no_spec_yml = validate_yaml_content(ConfigNamelistFile, "/path/to/some.nml")
     """
 
     specs: dict[str, Any] = field(default_factory=dict)
@@ -457,11 +457,11 @@ class ConfigIconTask(ConfigBaseTask):
     """
 
     plugin: ClassVar[Literal["icon"]] = "icon"
-    namelists: list[ConfigIconNamelist]
+    namelists: list[ConfigNamelistFile]
 
     @field_validator("namelists", mode="after")
     @classmethod
-    def check_nmls(cls, nmls: list[ConfigIconNamelist]) -> list[ConfigIconNamelist]:
+    def check_nmls(cls, nmls: list[ConfigNamelistFile]) -> list[ConfigNamelistFile]:
         # Make validator idempotent even if not used yet
         names = [nml.path.name for nml in nmls]
         if "icon_master.namelist" not in names:
