+++ tests/cases/large/config/config.yml
@@ -117,7 +117,9 @@ tasks:
       nodes: 40
       walltime: 23:59:59
       namelists:
-        - ./ICON/icon_master.namelist
+        - ./ICON/icon_master.namelist:
+            master_time_control_nml:
+              checkpointtimeintval: 'P2M'
         - ./ICON/NAMELIST_exclaim_ape_R02B04:
             parallel_nml:
               nproma: 96
