+++ tests/cases/large/data/ICON_namelists/icon_master.namelist_2026-09-01_00:00:00
@@ -6,8 +6,8 @@
 &master_time_control_nml
     calendar = 'proleptic gregorian'
     experimentstartdate = '2025-01-01T00:00:00Z'
-    restarttimeintval = 'PT1M'
-    checkpointtimeintval = 'PT1M'
+    restarttimeintval = 'P2M'
+    checkpointtimeintval = 'P2M'
     experimentstopdate = '2027-01-01T00:00:00Z'
 /
